requirements.txt
python-telegram-bot
nltk
tensorflow
snowballstemmer